Fix Ogmios connection error
This error would take place when the Ogmios service is started
with an invalid socket file. The Ogmios service is able to start and
receive connections but the handshake is never completed and the client
errors out when trying to send frames.

The fix was added to the underlying websocket_client library and this
commit simply updates Xogmios to point to the patched version.